Изменён тип переменных

+++ Pavlova_BEIS-20_LB2/Form1.cs
@@ -29,36 +29,36 @@ namespace Pavlova_BEIS_20_LB2
 
         private void button1_Click(object sender, EventArgs e)
         {
-            double a = Convert.ToDouble(textBox1.Text);
-            double b = Convert.ToDouble(textBox2.Text);
-            double c = a + b;
+            float a = (float)Convert.ToDouble(textBox1.Text);
+            float b = (float)Convert.ToDouble(textBox2.Text);
+            float c = a + b;
             textBox3.Text = c.ToString();
             label4.Text = "Сложение";
         }
 
         private void button2_Click(object sender, EventArgs e)
         {
-            double a = Convert.ToDouble(textBox1.Text);
-            double b = Convert.ToDouble(textBox2.Text);
-            double c = a - b;
+            float a = (float) Convert.ToDouble(textBox1.Text);
+            float b = (float) Convert.ToDouble(textBox2.Text);
+            float c = a - b;
             textBox3.Text = c.ToString();
             label4.Text = "Вычитание";
         }
 
         private void button3_Click(object sender, EventArgs e)
         {
-            double a = Convert.ToDouble(textBox1.Text);
-            double b = Convert.ToDouble(textBox2.Text);
-            double c = a * b;
+            float a = (float) Convert.ToDouble(textBox1.Text);
+            float b = (float) Convert.ToDouble(textBox2.Text);
+            float c = a * b;
             textBox3.Text = c.ToString();
             label4.Text = "Умножение";
         }
 
         private void button4_Click(object sender, EventArgs e)
         {
-            double a = Convert.ToDouble(textBox1.Text);
-            double b = Convert.ToDouble(textBox2.Text);
-            double c = a/b;
+            float a = (float) Convert.ToDouble(textBox1.Text);
+            float b = (float) Convert.ToDouble(textBox2.Text);
+            float c = a/b;
             textBox3.Text = c.ToString();
             label4.Text = "Деление";
         }
